+++ docs/release_notes.rst
@@ -14,6 +14,7 @@ Features in the 0.6.x series of releases are focused on expanding functionality 
 
     * Remove class ``BooleanExpressionTree`` in favor of working exclusively with instances of :class:`ExpressionTreeNode <tt.trees.tree_node.ExpressionTreeNode>`
     * Add :exc:`AlreadyConstrainedSymbolError <tt.errors.state.AlreadyConstrainedSymbolError>`
+    * Add :func:`sat_all <tt.satisfiability.picosat.sat_all>` to :mod:`picosat <tt.satisfiability.picosat>` interface
     * Add :func:`constrain <tt.expressions.bexpr.BooleanExpression.constrain>`, :func:`sat_one <tt.expressions.bexpr.BooleanExpression.sat_one>`, and :func:`sat_all <tt.expressions.bexpr.BooleanExpression.sat_all>` to :class:`BooleanExpression <tt.expressions.bexpr.BooleanExpression>`
     * Move the implementation logic of the :func:`to_cnf <tt.transformations.bexpr.to_cnf>` transformation to the :func:`to_cnf <tt.trees.tree_node.ExpressionTreeNode.to_cnf>` method of the :class:`ExpressionTreeNode <tt.trees.tree_node.ExpressionTreeNode>` class